Consolidate

+++ Diga.Core.Api.Win32/NmHdr.cs
@@ -8,6 +8,7 @@ namespace Diga.Core.Api.Win32
     {
         public IntPtr hwndFrom;
         public uint idFrom;
-        public long code;
+        
+        public IntPtr code;
     }
 }
